custom database exception handling

+++ src/main/java/com/jp/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.jp.dscommerce.services;
 import com.jp.dscommerce.dto.ProductDTO;
 import com.jp.dscommerce.entities.Product;
 import com.jp.dscommerce.repositories.ProductRepository;
+import com.jp.dscommerce.services.exceptions.DataBaseException;
 import com.jp.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 import java.util.Optional;
 
@@ -39,14 +43,27 @@ public class ProductService { //service devolve um DTO para o Controller e trata
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product product = productRepository.getReferenceById(id); //instancia um produto com a referência do id passado (não faz a busca no banco de dados)
-        copyDtoToEntity(product, dto);
-        product = productRepository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = productRepository.getReferenceById(id); //instancia um produto com a referência do id passado (não faz a busca no banco de dados)
+            copyDtoToEntity(product, dto);
+            product = productRepository.save(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("ID não encontrado");
+        }
     }
-    @Transactional
+
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("ID não encontrado");
+        }
+        try {
+            productRepository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DataBaseException("Falha de integridade referencial");
+        }
     }
 
     public void copyDtoToEntity(Product product, ProductDTO dto){
